+++ source/renderer/app/api/ada/adaTxFee.js
@@ -1,25 +1,15 @@
 // @flow
-import type { AdaTransactionFee } from './types';
-import { request } from './lib/request';
-
-export type AdaTxFeeParams = {
-  ca: string,
-  sender: string,
-  receiver: string,
-  amount: string,
-  // "groupingPolicy" - Spend everything from the address
-  // "OptimizeForSize" for no grouping
-  groupingPolicy: ?'OptimizeForSecurity' | 'OptimizeForSize',
-};
+import type { AdaTransactionFee, AdaTxFeeParams } from './types';
+import { request } from './lib/v1/request';
 
 export const adaTxFee = (
-  { ca, sender, receiver, amount, groupingPolicy }: AdaTxFeeParams
+  { ca, data }: AdaTxFeeParams
 ): Promise<AdaTransactionFee> => (
   request({
     hostname: 'localhost',
     method: 'POST',
-    path: `/api/txs/fee/${sender}/${receiver}/${amount}`,
+    path: '/api/v1/transactions/fees',
     port: 8090,
     ca,
-  }, {}, { groupingPolicy })
+  }, {}, data)
 );