+++ source/renderer/app/api/ada/newAdaPayment.js
@@ -1,27 +1,15 @@
 // @flow
-import type { AdaTransaction } from './types';
-import { request } from './lib/request';
-
-export type NewAdaPaymentParams = {
-  ca: string,
-  sender: string,
-  receiver: string,
-  amount: string,
-  password: ?string,
-  // "groupingPolicy" - Spend everything from the address
-  // "OptimizeForSize" for no grouping
-  groupingPolicy: ?'OptimizeForSecurity' | 'OptimizeForSize',
-};
-
+import type { AdaTransactionV1, AdaTransactionParams } from './types';
+import { request } from './lib/v1/request';
 
 export const newAdaPayment = (
-  { ca, sender, receiver, amount, groupingPolicy, password }: NewAdaPaymentParams
-): Promise<AdaTransaction> => (
+  { ca, data }: AdaTransactionParams
+): Promise<AdaTransactionV1> => (
   request({
     hostname: 'localhost',
     method: 'POST',
-    path: `/api/txs/payments/${sender}/${receiver}/${amount}`,
+    path: '/api/v1/transactions',
     port: 8090,
     ca,
-  }, { passphrase: password }, { groupingPolicy })
+  }, {}, data)
 );