+++ source/renderer/app/api/ada/newAdaWalletAddress.js
@@ -1,22 +1,22 @@
 // @flow
 import type { AdaAddress } from './types';
-import { request } from './lib/request';
+import { request } from './lib/v1/request';
 
 export type NewAdaWalletAddressParams = {
   ca: string,
-  password: ?string,
-  accountId: string,
+  spendingPassword?: string,
+  accountIndex: number,
+  walletId: string,
 };
 
 export const newAdaWalletAddress = (
-  { ca, password, accountId }: NewAdaWalletAddressParams
+  { ca, spendingPassword, accountIndex, walletId }: NewAdaWalletAddressParams
 ): Promise<AdaAddress> => (
   request({
     hostname: 'localhost',
     method: 'POST',
-    path: '/api/addresses',
+    path: '/api/v1/addresses',
     port: 8090,
     ca,
-  }, { passphrase: password }, accountId)
+  }, {}, { spendingPassword, accountIndex, walletId })
 );
-