+++ source/renderer/app/containers/wallet/WalletSendPage.js
@@ -40,8 +40,8 @@ export default class WalletSendPage extends Component<Props> {
         currencyMaxIntegerDigits={MAX_INTEGER_PLACES_IN_ADA}
         currencyMaxFractionalDigits={DECIMAL_PLACES_IN_ADA}
         validateAmount={validateAmount}
-        calculateTransactionFee={(receiver, amount) => (
-          calculateTransactionFee(activeWallet.id, receiver, amount)
+        calculateTransactionFee={(address: string, amount: number) => (
+          calculateTransactionFee({ walletId: activeWallet.id, address, amount })
         )}
         addressValidator={isValidAddress}
         isDialogOpen={uiDialogs.isOpen}